refactor(net): extract reusable PCP server (GatewayBackend trait)
Move the PCP MAP protocol core (RFC 6887 + HOSTNAME + PORT_SET handling,
response builders, constants) into core/src/net/pcp_server.rs, generic over
a GatewayBackend trait (add_forward/remove_forward/external_ipv4/
is_known_client/sni). Each gateway supplies its own sockets + forward
backend. TunnelContext implements GatewayBackend over its existing
nftables+PatchDb forwards, so the tunnel is behavior-preserving; this lets
StartWRT reuse the same PCP server with a UCI-firewall forward backend.
