+++ core/src/net/mod.rs
@@ -12,6 +12,7 @@ pub mod mdns;
 pub mod net_controller;
 pub mod pcp_hostname;
 pub mod pcp_portset;
+pub mod pcp_server;
 pub mod port_map;
 pub mod rfc2136;
 pub mod service_interface;
